Revamp the share/discussion parts.
Remove the hackernews/twitter ids, that's too much faff and inconsistent
as not all posts will have those links.  Instead go for a much simpler
set of 'share' links to submit a post to a bunch of different social
media sites.

Add OpenGraph metadata for Facebook/Google+.

Rename CSS id to postshare to reflect the change.